API

The main types below are exported. Remaining functions are not exported to avoid polluting the name space. Tip: To simplify calling the following functions it is possible to write import FerriteProblems as FP as is done in the examples.

Main types

FerriteProblems.FerriteProblem — Type

FerriteProblem(def::FEDefinition, post=nothing, io=nothing)

Create a FerriteProblem from def. Postprocessing can be added as post, see FESolvers.postprocess!. File input/output using FerriteIO can be added with io. It is possible to give the folder where to save the output (i.e. io::String), or to construct FerriteIO with more options directly.

FerriteProblems.FEDefinition — Type

Create the FEDefinition which can later be used to create a complete FerriteProblem.

FEDefinition(;dh, ch, material, cellvalues, 
    nh=NeumannHandler(dh),
    convergence_criterion=AbsoluteResidual(), 
    initial_conditions=NamedTuple(),
    user_data=nothing, 
    autodiffbuffer=Val(false), threading=Val(false)
    )

This first definition sets up a single domain simulation (same material, cellvalues, interpolations, etc. everywhere).

FEDefinition(domains::Vector{<:AssemblyDomain}}; ch, 
    nh=NeumannHandler(dh),
    convergence_criterion=AbsoluteResidual(), 
    initial_conditions=NamedTuple(), 
    autodiffbuffer=Val(false), threading=Val(false)
    )

This second definition sets up a multi-domain simulation by using FerriteAssembly.jl's AssemblyDomain. The dh used for the default NeumannHandler is extracted from domains.

Common inputs

For both constructors cases, the following inputs can be supplied

ch: The constraint handler, ConstraintHandler (Ferrite.jl)

nh: The neumann bc handler, NeumannHandler (FerriteNeumann.jl)

initial_conditions: NamedTuple with a function f(x) for each field that has a nonzero initial condition. Used by the Ferrite.jl's apply_analytical! function. Example: initial_conditions = (u = x -> Vec((x[1]/10, 0.0)), p = x -> -100*x[2]). For fields not given here, the initial condition is zeros everywhere.

convergence_criterion: Determines how to calculate the convergence measure including scaling. See ConvergenceCriterion

autodiffbuffer: Should FerriteAssembly.jl's AutoDiffCellBuffer be used? This will make the assembly faster if automatic differentiation is used, and can also be used without automatic differentiation (but with a slight extra computational overhead) Bool input can also be given (will be converted internally to Val)

threading: Should threading be used? Bool input can also be given (will be converted internally to Val)

FerriteProblems.FerriteIO — Type

FerriteIO(
    folder::String, def::FEDefinition, post=nothing; 
    def_file="FEDefinition.jld2", 
    postfile="FEPost.jld2",
    T=Float64, 
    nsteps_per_file=typemax(Int), 
    switchsize=Inf
    )

Constructor for creating a FerriteIO when simulating.

FerriteIO(filename::String)

Constructor for reading a FerriteIO that was saved during a simulation

Convergence criteria

In normal usage, the following convergence criteria can be used

FerriteProblems.ConvergenceCriterion — Type

ConvergenceCriterion

The abstract type ConvergenceCriterion is the supertype for all convergence criteria.

FerriteProblems.AbsoluteResidual — Type

AbsoluteResidual()

The default convergence criterion that calculates the convergence measure as √(sum([r[i]^2 for i ∈ free dofs]) without any scaling.

FerriteProblems.RelativeResidualElementScaling — Type

RelativeResidualElementScaling(p, minfactors::Union{AbstractFloat,NamedTuple}=eps())

Use Ferriteassembly.ElementResidualScaling with the exponent p to calculate the scaling for each field individually, based on the L2-norm of each cell's residual. To avoid issues when all cells have zero residual (e.g. in the first time step), supply minfactors as the minimum scaling factor. The convergence measure is calculated with the following pseudo-code

val = 0.0
for field in Ferrite.getfieldnames(dh)
    factor = max(element_residual_scaling[field], minfactors[field])
    dofs = free_field_dofs[field]   # Get the non-constrained dofs for `field`
    val += sum(i->(r[i]/factor)^2, dofs)
end
return √val

where the same minfactor is used for all fields if only a scalar value is given.

To create custom convergence criteria, the following functions may require overloading.

FerriteProblems.TolScaling — Type

TolScaling(criterion::ConvergenceCriterion, def::FEDefinition)

The TolScaling type contains the criterion that determines how to scale the residuals to determine convergence. The constructor is specialized on typeof(criterion), creating the following fields:

assemscaling: scaling to be used by FerriteAssembly to give potential scaling contribution based on each element's residual

buffer: Used to pre-calculate values, such as dof-ranges for each field that is used when calculating the convergence measure.

FESolvers.calculate_convergence_measure — Function

FESolvers.calculate_convergence_measure(::ConvergenceCriterion, scaling::TolScaling, r::Vector, Δa, iter, p::FerriteProblem)

Specialize this function for each ConvergenceCriterion. Note this function is called via the definition

calculate_convergence_measure(ts::TolScaling, args...) = calculate_convergence_measure(ts.criterion, ts, args...)

FerriteProblems.make_assemscaling — Function

make_assemscaling(criterion, def)

Create the scaling for use in FerriteAssembly if required by the given criterion. The default, if not overloaded, returns nothing.

Access functions

FerriteProblems.getdh — Function

FerriteProblems.getdh(p::FerriteProblem)

Get dh::Ferrite.AbstractDofHandler from the FerriteProblem

FerriteProblems.getch — Function

FerriteProblems.getch(p::FerriteProblem)

Get the ConstraintHandler from the FerriteProblem

FerriteProblems.getnh — Function

FerriteProblems.getnh(p::FerriteProblem)

Get the NeumannHandler from the FerriteProblem

FerriteAssembly.get_material — Function

FerriteProblems.get_material(p::FerriteProblem)
FerriteProblems.get_material(p::FerriteProblem, domain_name::String)

Get the material in p. For multiple domains, it is necessary to give the domain_name for where to get the material. Note that this is type-unstable and should be avoided in performance-critical code sections. This function belongs to FerriteAssembly.jl, but can be accessed via FerriteProblems.get_material.

FESolvers.getjacobian — Function

FerriteProblems.getjacobian(p::FerriteProblem)

Get the current jacobian matrix from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getjacobian

FESolvers.getunknowns — Function

FerriteProblems.getunknowns(p::FerriteProblem)

Get the current vector of unknowns from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getunknowns

FESolvers.getresidual — Function

FerriteProblems.getresidual(p::FerriteProblem)

Get the current residual vector from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getresidual

FerriteProblems.getneumannforce — Function

FerriteProblems.getneumannforce(p::FerriteProblem)

Get the current external force vector caused by Neumann boundary conditions. Note that this vector does not include external forces added during the cell assembly; only forces added with the NeumannHandler

FerriteProblems.getoldunknowns — Function

FerriteProblems.getoldunknowns(p::FerriteProblem)

Get the vector of unknowns from the previously converged step

FerriteProblems.getstate — Function

FerriteProblems.getstate(p::FerriteProblem)

Get the current state variables

FerriteProblems.getoldstate — Function

FerriteProblems.getoldstate(p::FerriteProblem)

Get the state variables from the previously converged step

FerriteProblems.gettime — Function

FerriteProblems.gettime(p::FerriteProblem)

Get the current time

FerriteProblems.getoldtime — Function

FerriteProblems.getoldtime(p::FerriteProblem)

Get time of the previous converged step

Saving and loading data

FESolvers.postprocess! — Function

FESolvers.postprocess!(p::FerriteProblem, step, solver)

When FESolvers call this function for p::FerriteProblem, the following function

FESolvers.postprocess!(post, p::FerriteProblem, step, solver)

is called where post=p.post (unless you define a different override). This allows you to easily define the dispatch on your postprocessing type as FESolvers.postprocess!(post::MyPostType, p, step, solver) Note that the solver input argument is required, but can be accounted for by defining, e.g. FESolvers.postprocess!(post::MyPostType, p, step, args...)

FerriteProblems.close_postprocessing — Function

close_postprocessing(post::MyPostType, p::FerriteProblem)

This function is called to close any open files manually created during the postprocessing with the custom postprocessing type MyPostType. Note that the file streams in p.io::FerriteIO are automatically closed and don't require any special handling.

FerriteProblems.addstep! — Function

FerriteProblems.addstep!(io::FerriteIO, p::FerriteProblem)

Add a new step to be saved by io at the time gettime(p) Must be called before adding any new data

FerriteProblems.gettimedata — Function

gettimedata(io::FerriteIO)
gettimedata(io::FerriteIO, step)

FerriteProblems.savedofdata! — Function

FerriteProblems.savedofdata!(io::FerriteIO, vals, dt_order=0, field="dof")

Save data pertaining to each degree of freedom. Use a different field than "dof"` to save data located at each dof, but not the actual dof values (e.g. the residual vector)

FerriteProblems.getdofdata — Function

FerriteProblems.getdofdata(io::FerriteIO, step, field="dof"; dt_order=0)

Get the data saved by FerriteProblems.savedofdata! in step for field.

FerriteProblems.savenodedata! — Function

FerriteProblems.savenodedata!(io::FerriteIO, vals, field, dt_order=0)

Save data located at each node. By convention this should be indexed by the node numbers in the grid. (E.g. a Vector for all nodes or a Dict{Int} with keys the node numbers)

FerriteProblems.getnodedata — Function

FerriteProblems.getnodedata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.savenodedata! in step for field.

FerriteProblems.savecelldata! — Function

FerriteProblems.savecelldata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each cell. By convention this should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices)

FerriteProblems.getcelldata — Function

FerriteProblems.getcelldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.savecelldata! in step for field.

FerriteProblems.saveipdata! — Function

FerriteProblems.saveipdata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each integration point in cells in the grid. By convention the data for each cell should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices) Note that it is on the user to know how the integration points are numbered, i.e. which QuadratureRule that was used.

FerriteProblems.getipdata — Function

FerriteProblems.getipdata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.saveipdata! in step for field.

FerriteProblems.saveglobaldata! — Function

FerriteProblems.saveglobaldata!(io::FerriteIO, vals, field, dt_order=0)

Save data that is global to the entire simulation, i.e. global quantites such as reaction forces, total dissipation, etc.

FerriteProblems.getglobaldata — Function

FerriteProblems.getglobaldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.saveglobaldata! in step for field.

FerriteProblems.getdef — Function

FerriteProblems.getdef(io::FerriteIO)

Load the FEDefinition from the results saved by io

FerriteProblems.getpost — Function

FerriteProblems.getpost(io::FerriteIO)

Load the user defined post from the results saved by io