API

The main types below are exported. Remaining functions are not exported to avoid polluting the name space. Tip: To simplify calling the following functions it is possible to write import FerriteProblems as FP as is done in the examples.

Main types

FerriteProblems.FerriteProblem — Type

FerriteProblem(def::FEDefinition, post, buf::FEBuffer, io::[FerriteIO])

The main problem type that holds all variables to solve a particular problem using FESolvers

FerriteProblems.FEDefinition — Type

FEDefinition(dh, ch, nh, cv, m, bl, initialstate, ic)
FEDefinition(;
    dh, ch, cv, m,
    nh=NeumannHandler(dh), bl=nothing, ic=(), 
    initialstate=_create_states(dh,m,cv,ic),
    cc=AbsoluteResidual(), colors=nothing,
    )

The definition of the FE-problem, this data describes the finite element problem as a function of time. For how long to run and how to step through time is typically done by FESolvers.jl.

Fields/inputs

See more information below for items marked with *

dh: The dof handler, Union{DofHandler,MixedDofHandler} (Ferrite.jl)

ch: The constraint handler, ConstraintHandler (Ferrite.jl)

nh: The neumann bc handler, NeumannHandler (FerriteNeumann.jl)

cv: The cellvalues, e.g. CellValues (*)

m: The material definition, user defined type - passed into element. (*)

bl: Source term/body load, user defined type - available from the element routine (*)

ic: Initial conditions. NamedTuple with a function f(x) for each field that has a nonzero initial condition. Used by the Ferrite.jl's apply_analytical! function (included here if Ferrite version too old).

initialstate: The initial state variables for each cell in the grid. By default, this is automatically created by FerriteAssembly.create_states with customization provided by overloading FerriteAssembly.create_cell_state and is not necessary to supply here.

cc::ConvergenceCriterion: Determines how to calculate the convergence measure including scaling

colors: Supply to use threaded assembly, can be created by Ferrite.create_coloring

MixedDofHandler

When the MixedDofHandler is used, we have an outer loop over each of its FieldHandlers. Therefore, cv, m, and bl may be provided for each of these fields by providing them as a tuple. Otherwise, they are duplicated for each field (by reference)

initialstate must be a tuple containing Dict{Int} with states for the cells in the relevant FieldHandler's cellset.

Multiple CellValues for each element In coupled problems, each element might require multiple CellValues. If the same should be used for all fields (even if only one FieldHandler), a NamedTuple is the easiest (the alternative is a Tuple of Tuples). If different multiple CellValues should be used for each field, then pass as Tuple of NamedTuple/Tuple

Multiple materials

Note: The following setup is not needed if you have one material for each FieldHandler when using a MixedDofHandler

If you want to have different materials on different parts of the grid, you should first define those cellsets, and then define each material as a value of m::Dict{String} with the name of the cellset as key.

You may do the same for bl and cv. Otherwise, the same values will be used for every cellset.

Combine with MixedDofHandler: Follow the required input for MixedDofHandler, but wrap items (at least the material and initialstate) in Dicts as described above

State variables: initialstate::Dict{String} must mirror the datastructure of m::Dict{String}: For each key in m::Dict{String}, initialstate[key] holds a collection of states for the cells in getcellset(dh,key).

If MixedDofHandler, then initialstate[key]::NTuple{N,<:Dict{Int}}, and if DofHandler, then initialstate[key]::Dict{Int}

FerriteProblems.FerriteIO — Type

FerriteIO(
    folder::String, def::FEDefinition, post=nothing; 
    def_file="FEDefinition.jld2", 
    postfile="FEPost.jld2",
    T=Float64, 
    nsteps_per_file=typemax(Int), 
    switchsize=Inf
    )

Constructor for creating a FerriteIO when simulating.

FerriteIO(filename::String)

Constructor for reading a FerriteIO that was saved during a simulation

Setting up simulation

FerriteProblems.cellbuffertype — Function

cellbuffertype(def)

By default, cellbuffertype(::Any) = CellBuffer. However, if you are using the automatic differentiation, much better assembly speed can be achieved by defining

FerriteProblems.cellbuffertype(::FEDefinition) = AutoDiffCellBuffer

In this case, any defined element_routine! should use getCellBuffer(buffer) to get a CellBuffer instead of AutoDiffCellBuffer. This is not necessary in element_residual!

FerriteProblems.allocate_material_cache — Function

FerriteProblems.allocate_material_cache(material, cellvalues)

In case the material requires a cache to be available during the element routine, this function can be overloaded for the specific material to define such a cache to be included in the FerriteAssembly.CellBuffer

Convergence criteria

In normal usage, the following convergence criteria can be used

FerriteProblems.ConvergenceCriterion — Type

ConvergenceCriterion

The abstract type ConvergenceCriterion is the supertype for all convergence criteria.

FerriteProblems.AbsoluteResidual — Type

AbsoluteResidual()

The default convergence criterion that calculates the convergence measure as √(sum([r[i]^2 for i ∈ free dofs]) without any scaling.

FerriteProblems.RelativeResidualElementScaling — Type

RelativeResidualElementScaling(p, minfactors::Union{AbstractFloat,NamedTuple}=eps())

Use Ferriteassembly.ElementResidualScaling with the exponent p to calculate the scaling for each field individually, based on the L2-norm of each cell's residual. To avoid issues when all cells have zero residual (e.g. in the first time step), supply minfactors as the minimum scaling factor. The convergence measure is calculated with the following pseudo-code

val = 0.0
for field in Ferrite.getfieldnames(dh)
    factor = max(element_residual_scaling[field], minfactors[field])
    dofs = free_field_dofs[field]   # Get the non-constrained dofs for `field`
    val += sum(i->(r[i]/factor)^2, dofs)
end
return √val

where the same minfactor is used for all fields if only a scalar value is given.

To create custom convergence criteria, the following functions may require overloading.

FerriteProblems.TolScaling — Type

TolScaling(criterion::ConvergenceCriterion, def::FEDefinition)

The TolScaling type contains the criterion that determines how to scale the residuals to determine convergence. The constructor is specialized on typeof(criterion), creating the following fields:

assemscaling: Used in FerriteAssembly.doassemble! to give potential scaling contribution based on each element's residual

buffer: Used to pre-calculate values, such as dof-ranges for each field that is used when calculating the convergence measure.

FerriteProblems.make_assemscaling — Function

make_assemscaling(criterion, def, threaded::Val{false})

Create the actual assemscaling for the given criterion.

make_assemscaling(criterion, def, threaded::Val{true})

If threaded::Val{true}, then call make_assemscaling(criterion, def, ::Val{false}) for each thread to support threaded assembly. Note that the function has a definition make_assemscaling(criterion,def)=make_assemscaling(criterion,def,dothreaded(def)) to simplify calling it. If required, a criterion should overload make_assemscaling(criterion,def,::Val{false}). The default, if not overloaded, returns nothing. That suffices if nothing needs to be calculated for each cell during assembly.

FESolvers.calculate_convergence_measure — Function

FESolvers.calculate_convergence_measure(::ConvergenceCriterion, scaling::TolScaling, r::Vector, Δa, iter, p::FerriteProblem)

Specialize this function for each ConvergenceCriterion. Note this function is called via the definition

calculate_convergence_measure(ts::TolScaling, args...) = calculate_convergence_measure(ts.criterion, ts, args...)

Access functions

FerriteProblems.getdh — Function

FerriteProblems.getdh(p::FerriteProblem)

Get dh::Ferrite.AbstractDofHandler from the FerriteProblem

FerriteProblems.getch — Function

FerriteProblems.getch(p::FerriteProblem)

Get the ConstraintHandler from the FerriteProblem

FerriteProblems.getnh — Function

FerriteProblems.getnh(p::FerriteProblem)

Get the NeumannHandler from the FerriteProblem

FerriteProblems.getcv — Function

FerriteProblems.getcv(p::FerriteProblem)

Get the cell values from the FerriteProblem. Note that this could also be a Tuple or NamedTuple depending on what was initially given to FEDefinition

FerriteProblems.getmaterial — Function

FerriteProblems.getmaterial(p::FerriteProblem)

Get the material from the FerriteProblem

FerriteProblems.getbodyload — Function

FerriteProblems.getbodyload(p::FerriteProblem)

Get the bodyload given to the FerriteProblem

FESolvers.getjacobian — Function

FerriteProblems.getjacobian(p::FerriteProblem)

Get the current jacobian matrix from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getjacobian

FESolvers.getunknowns — Function

FerriteProblems.getunknowns(p::FerriteProblem)

Get the current vector of unknowns from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getunknowns

FESolvers.getresidual — Function

FerriteProblems.getresidual(p::FerriteProblem)

Get the current residual vector from p. Note that this function belongs to FESolvers.jl, but can be accessed via FerriteProblems.getresidual

FerriteProblems.getneumannforce — Function

FerriteProblems.getneumannforce(p::FerriteProblem)

Get the current external force vector caused by Neumann boundary conditions. Note that this vector does not include external forces added during the cell assembly; only forces added with the NeumannHandler

FerriteProblems.getoldunknowns — Function

FerriteProblems.getoldunknowns(p::FerriteProblem)

Get the vector of unknowns from the previously converged step

FerriteProblems.getstate — Function

FerriteProblems.getstate(p::FerriteProblem)

Get the current state variables

FerriteProblems.getoldstate — Function

FerriteProblems.getoldstate(p::FerriteProblem)

Get the state variables from the previously converged step

FerriteProblems.gettime — Function

FerriteProblems.gettime(p::FerriteProblem)

Get the current time

FerriteProblems.getoldtime — Function

FerriteProblems.getoldtime(p::FerriteProblem)

Get time of the previous converged step

Saving and loading data

FESolvers.postprocess! — Function

FESolvers.postprocess!(p::FerriteProblem, step, solver)

When FESolvers call this function for p::FerriteProblem, the following function

FESolvers.postprocess!(post, p::FerriteProblem, step, solver)

is called where post=p.post (unless you define a different override). This allows you to easily define the dispatch on your postprocessing type as FESolvers.postprocess!(post::MyPostType, p, step, solver) Note that the solver input argument is required, but can be accounted for by defining, e.g. FESolvers.postprocess!(post::MyPostType, p, step, args...)

FerriteProblems.close_postprocessing — Function

close_postprocessing(post::MyPostType, p::FerriteProblem)

This function is called to close any open files manually created during the postprocessing with the custom postprocessing type MyPostType. Note that the file streams in p.io::FerriteIO are automatically closed and don't require any special handling.

FerriteProblems.addstep! — Function

FerriteProblems.addstep!(io::FerriteIO, p::FerriteProblem)

Add a new step to be saved by io at the time gettime(p) Must be called before adding any new data

FerriteProblems.gettimedata — Function

gettimedata(io::FerriteIO)
gettimedata(io::FerriteIO, step)

FerriteProblems.savedofdata! — Function

FerriteProblems.savedofdata!(io::FerriteIO, vals, dt_order=0, field="dof")

Save data pertaining to each degree of freedom. Use a different field than "dof"` to save data located at each dof, but not the actual dof values (e.g. the residual vector)

FerriteProblems.getdofdata — Function

FerriteProblems.getdofdata(io::FerriteIO, step, field="dof"; dt_order=0)

Get the data saved by FerriteProblems.savedofdata! in step for field.

FerriteProblems.savenodedata! — Function

FerriteProblems.savenodedata!(io::FerriteIO, vals, field, dt_order=0)

Save data located at each node. By convention this should be indexed by the node numbers in the grid. (E.g. a Vector for all nodes or a Dict{Int} with keys the node numbers)

FerriteProblems.getnodedata — Function

FerriteProblems.getnodedata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.savenodedata! in step for field.

FerriteProblems.savecelldata! — Function

FerriteProblems.savecelldata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each cell. By convention this should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices)

FerriteProblems.getcelldata — Function

FerriteProblems.getcelldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.savecelldata! in step for field.

FerriteProblems.saveipdata! — Function

FerriteProblems.saveipdata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each integration point in cells in the grid. By convention the data for each cell should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices) Note that it is on the user to know how the integration points are numbered, i.e. which QuadratureRule that was used.

FerriteProblems.getipdata — Function

FerriteProblems.getipdata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.saveipdata! in step for field.

FerriteProblems.saveglobaldata! — Function

FerriteProblems.saveglobaldata!(io::FerriteIO, vals, field, dt_order=0)

Save data that is global to the entire simulation, i.e. global quantites such as reaction forces, total dissipation, etc.

FerriteProblems.getglobaldata — Function

FerriteProblems.getglobaldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by FerriteProblems.saveglobaldata! in step for field.

FerriteProblems.getdef — Function

FerriteProblems.getdef(io::FerriteIO)

Load the FEDefinition from the results saved by io

FerriteProblems.getpost — Function

FerriteProblems.getpost(io::FerriteIO)

Load the user defined post from the results saved by io