API

Main types and running simulations

FerriteProblems.FerriteProblem — Type

FerriteProblem(def::FEDefinition, post=nothing, io=nothing)

Constructor that makes the minimum required to run simulation. Optional postprocessing and io::FerriteIO if desired.

FerriteProblem(def::FEDefinition, post, savefolder::String)

Constructor with automatic generation of FerriteIO by just specifying in which folder data should be saved.

FerriteProblems.FEDefinition — Type

FEDefinition(dh, ch, nh, cv, m, bl, initialstate, ic)
FEDefinition(;dh, ch, cv, m,
    nh=NeumannHandler(dh), bl=nothing,
    initialstate=create_empty_states(dh,m),
    ic=(),
    )

The definition of the FE-problem, this data describes the finite element problem as a function of time. For how long to run and how to step through time is typically done by FESolvers.jl.

Fields/inputs

See more information below for items marked with *

dh: The dof handler, Union{DofHandler,MixedDofHandler} (Ferrite.jl)

ch: The constraint handler, ConstraintHandler (Ferrite.jl)

nh: The neumann bc handler, NeumannHandler (FerriteNeumann.jl)

cv: The cellvalues, e.g. CellValues (*)

m: The material definition, user defined type - passed into element. (*)

bl: Source term/body load, user defined type - available from the element routine (*)

initialstate: The initial state variables for each cell in the grid, can created by FerriteAssembly.create_states (*)

ic: Initial conditions. NamedTuple with a function f(x) for each field that has a nonzero initial condition. Used by the initial_conditions! function.

MixedDofHandler

When the MixedDofHandler is used, we have an outer loop over each of its FieldHandlers. Therefore, cv, m, and bl may be provided for each of these fields by providing them as a tuple. Otherwise, they are duplicated for each field (by reference)

initialstate must be a tuple containing Dict{Int} with states for the cells in the relevant FieldHandler's cellset.

Multiple CellValues for each element In coupled problems, each element might require multiple CellValues. If the same should be used for all fields (even if only one FieldHandler), a NamedTuple is the easiest (the alternative is a Tuple of Tuples). If different multiple CellValues should be used for each field, then pass as Tuple of NamedTuple/Tuple

Multiple materials

Note: The following setup is not needed if you have one material for each FieldHandler when using a MixedDofHandler

If you want to have different materials on different parts of the grid, you should first define those cellsets, and then define each material as a value of m::Dict{String} with the name of the cellset as key.

You may do the same for bl and cv. Otherwise, the same values will be used for every cellset.

Combine with MixedDofHandler: Follow the required input for MixedDofHandler, but wrap items (at least the material and initialstate) in Dicts as described above

State variables: initialstate::Dict{String} must mirror the datastructure of m::Dict{String}: For each key in m::Dict{String}, initialstate[key] holds a collection of states for the cells in getcellset(dh,key).

If MixedDofHandler, then initialstate[key]::NTuple{N,<:Dict{Int}}, and if DofHandler, then initialstate[key]::Dict{Int}

FerriteProblems.FerriteIO — Type

FerriteIO(
    folder::String, def::FEDefinition, post=nothing; 
    def_file="FEDefinition.jld2", 
    postfile="FEPost.jld2",
    T=Float64, 
    nsteps_per_file=typemax(Int), 
    switchsize=Inf
    )

Constructor for creating a FerriteIO when simulating.

FerriteIO(filename::String)

Constructor for reading a FerriteIO that was saved during a simulation

FerriteProblems.safesolve — Function

safesolve(solver, def::FEDefinition, args...; kwargs...)

Starts by creating FerriteProblem from def and args/kwargs and then wraps the call to FESolvers.solve_problem! in try ... finally to ensure that close_problem is called even in the case of no convergence.

FerriteProblems.close_problem — Function

close_problem(p::FerriteProblem)

Method for closing all open files before ending the simulation

Access functions

FerriteProblems.getdh — Function

FP.getdh(p::FerriteProblem)

Get dh::Ferrite.AbstractDofHandler from the FEDefinition

FerriteProblems.getch — Function

FP.getch(p::FerriteProblem)

Get the ConstraintHandler from the FEDefinition

FerriteProblems.getnh — Function

FP.getnh(p::FerriteProblem)

Get the NeumannHandler from the FEDefinition

FerriteProblems.getcv — Function

FP.getcv(p::FerriteProblem)

Get the cell values from the FEDefinition. Note that this could also be a Tuple or NamedTuple depending on what was initially given to FEDefinition

FerriteProblems.getmaterial — Function

FP.getmaterial(p::FerriteProblem)

Get the material from the FEDefinition

FerriteProblems.getbodyload — Function

FP.getbodyload(p::FerriteProblem)

Get the bodyload given to the FEDefinition

FESolvers.getjacobian — Function

FP.getjacobian(p::FerriteProblem)

Get the current jacobian matrix from p. Note that this function belongs to FESolvers.jl, but can be accessed via FP.getjacobian

FESolvers.getunknowns — Function

FP.getunknowns(p::FerriteProblem)

Get the current vector of unknowns from p. Note that this function belongs to FESolvers.jl, but can be accessed via FP.getunknowns

FESolvers.getresidual — Function

FP.getresidual(p::FerriteProblem)

Get the current residual vector from p. Note that this function belongs to FESolvers.jl, but can be accessed via FP.getresidual

FerriteProblems.getneumannforce — Function

FP.getneumannforce(p::FerriteProblem)

Get the current external force vector caused by Neumann boundary conditions. Note that this vector does not include external forces added during the cell assembly; only forces added with the NeumannHandler

FerriteProblems.getoldunknowns — Function

FP.getoldunknowns(p::FerriteProblem)

Get the vector of unknowns from the previously converged step

FerriteProblems.getstate — Function

FP.getstate(p::FerriteProblem)

Get the current state variables

FerriteProblems.getoldstate — Function

FP.getoldstate(p::FerriteProblem)

Get the state variables from the previously converged step

FerriteProblems.gettime — Function

FP.gettime(p::FerriteProblem)

Get the current time

FerriteProblems.getoldtime — Function

FP.getoldtime(p::FerriteProblem)

Get time of the previous converged step

Special cases

FerriteProblems.allocate_material_cache — Function

FP.allocate_material_cache(material)

In case the material requires a cache to be available during the element routine, this function can be overloaded for the specific material to define such a cache to be included in the FerriteAssembly.CellBuffer

Saving and loading data

FESolvers.postprocess! — Function

FESolvers.postprocess!(p::FerriteProblem, step, solver)

When FESolvers call this function for p::FerriteProblem, the following function

FESolvers.postprocess!(post, p::FerriteProblem, step, solver)

is called where post=p.post (unless you define a different override). This allows you to easily define the dispatch on your postprocessing type as FESolvers.postprocess!(post::MyPostType, p, step, solver)

FerriteProblems.addstep! — Function

FP.addstep!(io::FerriteIO, p::FerriteProblem)

Add a new step to be saved by io at the time gettime(p) Must be called before adding any new data

FerriteProblems.gettimedata — Function

gettimedata(io::FerriteIO)
gettimedata(io::FerriteIO, step)

FerriteProblems.savedofdata! — Function

FP.savedofdata!(io::FerriteIO, vals, dt_order=0, field="dof")

Save data pertaining to each degree of freedom. Use a different field than "dof"` to save data located at each dof, but not the actual dof values (e.g. the residual vector)

FerriteProblems.getdofdata — Function

FP.getdofdata(io::FerriteIO, step, field="dof"; dt_order=0)

Get the data saved by savedofdata! in step for field.

FerriteProblems.savenodedata! — Function

FP.savenodedata!(io::FerriteIO, vals, field, dt_order=0)

Save data located at each node. By convention this should be indexed by the node numbers in the grid. (E.g. a Vector for all nodes or a Dict{Int} with keys the node numbers)

FerriteProblems.getnodedata — Function

FP.getnodedata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by savenodedata! in step for field.

FerriteProblems.savecelldata! — Function

FP.savecelldata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each cell. By convention this should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices)

FerriteProblems.getcelldata — Function

FP.getcelldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by savecelldata! in step for field.

FerriteProblems.saveipdata! — Function

FP.saveipdata!(io::FerriteIO, vals, field, dt_order=0)

Save data for each integration point in cells in the grid. By convention the data for each cell should be indexed by the cell numbers in the grid. (E.g. a Vector for all cells or a Dict{Int} with keys the cell indices) Note that it is on the user to know how the integration points are numbered, i.e. which QuadratureRule that was used.

FerriteProblems.getipdata — Function

FP.getipdata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by saveipdata! in step for field.

FerriteProblems.saveglobaldata! — Function

FP.saveglobaldata!(io::FerriteIO, vals, field, dt_order=0)

Save data that is global to the entire simulation, i.e. global quantites such as reaction forces, total dissipation, etc.

FerriteProblems.getglobaldata — Function

FP.getglobaldata(io::FerriteIO, step, field; dt_order=0)

Get the data saved by saveglobaldata! in step for field.

FerriteProblems.getdef — Function

FP.getdef(io::FerriteIO)

Load the FEDefinition from the results saved by io

FerriteProblems.getpost — Function

FP.getpost(io::FerriteIO)

Load the user defined post from the results saved by io