Internal API

Note: These functions are not considered "documented" and changes here are therefore not considered breaking.

FerriteProblems.FEBuffer — Type

FEBuffer(K,x,r,f,xold,cellbuffer,state,old_state,time,old_time)

A buffer to hold all values that are required to simulate, but that are uniqely defined from the simulation definition

FerriteProblems.getcellbuffer — Function

FP.getcellbuffer(p::FerriteProblem)

Get the cell buffers used during the assembly.

FerriteProblems.update_unknowns! — Function

FP.update_unknowns!(p::FerriteProblem)

Update the vector of "old" unknowns to the values of the current vector of unknowns

FerriteProblems.copy_states! — Function

FP.copy_states!(to, from)

Perform a "deepcopy!" of states in from into to.

FerriteProblems.update_states! — Function

FP.update_states!(p::FerriteProblem)

Update the "old" state variables to the current values. Called after convergence

FerriteProblems.reset_states! — Function

FP.reset_states!(p::FerriteProblem)

Reset the current state variables to the old state values. Called after each solution iteration unless the solution has converged.

FerriteProblems.settime! — Function

FP.settime!(p::FerriteProblem, new_time)

Set the current time to new_time Called when starting a new step (or when attempting the same step number with a new time increment)

FerriteProblems.update_time! — Function

FP.update_time!(p::FerriteProblem)

Update the old time to the current time. Called after convergence

FerriteProblems.create_empty_states — Function

FP.create_empty_states(dh, material)

Used to create empty states in case state variables aren't used in the simulation by calling FerriteAssembly.create_states without any input (returning nothing for each cell)

FerriteProblems.filepath — Function

filepath(io::FerriteIO, args...) = joinpath(io.folder[], args...)

Get the path of a file relative io's folder

FerriteProblems.datafilepath — Function

datafilepath(io::FerriteIO, num=length(io.datafiles))

Get the path of the data file number num in io.datafiles

FerriteProblems.new_file! — Function

new_file!

FerriteProblems.new_file_if_needed! — Function

new_file_if_needed!

FerriteProblems.update_currentsize! — Function

update_currentsize!

FerriteProblems.getfilenumber — Function

getfilenumber

FerriteProblems.open_if_needed! — Function

open_if_needed!

FerriteProblems.checkkey — Function

checkkey

FerriteProblems.savedata! — Function

savedata!

FerriteProblems.getdata — Function

getdata

FerriteProblems.get_dof2node — Function

get_dof2node

FerriteProblems.initial_conditions! — Function

initial_conditions!(
    a::AbstractVector, dh::DofHandler, field::Symbol, f::Function, 
    cellset=1:getncells(dh.grid),
    ip_geo=_default_interpolation(dh)
    )

Apply initial conditions using the function f