Internal API

Note: These functions are not considered "documented" and changes here are therefore not considered breaking.

FerriteProblems.FEBuffer — Type

FEBuffer(K,x,r,f,xold,cellbuffer,state,old_state,time,old_time)

A buffer to hold all values that are required to simulate, but that are uniqely defined from the simulation definition

FerriteProblems.get_tolerance_scaling — Function

FerriteProblems.get_tolerance_scaling(p::FerriteProblem)

Get the TolScaling type that controls the convergence measure to be compared with the solver's tolerance

FerriteProblems.update_unknowns! — Function

FerriteProblems.update_unknowns!(p::FerriteProblem)

Update the vector of "old" unknowns to the values of the current vector of unknowns

FerriteProblems.set_time! — Function

FerriteProblems.settime!(p::FerriteProblem, new_time)

Set the current time to new_time Called when starting a new step (or when attempting the same step number with a new time increment)

FerriteProblems.update_time! — Function

FerriteProblems.update_time!(p::FerriteProblem)

Update the old time to the current time. Called after convergence

FerriteProblems.filepath — Function

filepath(io::FerriteIO, args...) = joinpath(io.folder, args...)

Get the path of a file relative io's folder

FerriteProblems.datafilepath — Function

datafilepath(io::FerriteIO, num=length(io.datafiles))

Get the path of the data file number num in io.datafiles

FerriteProblems.new_file! — Function

new_file!

FerriteProblems.new_file_if_needed! — Function

new_file_if_needed!

FerriteProblems.update_currentsize! — Function

update_currentsize!

FerriteProblems.getfilenumber — Function

getfilenumber

FerriteProblems.open_if_needed! — Function

open_if_needed!

FerriteProblems.close_io — Function

close_io(io::FerriteIO, post)

Close the currently open file in io, then the postprocessing struct to a jld2 file (if not post != nothing), before finally saving the current io object to a .jld2 file

FerriteProblems.checkkey — Function

checkkey

FerriteProblems.savedata! — Function

savedata!

FerriteProblems.getdata — Function

getdata