Internal API

Note: These functions are not considered "documented" and changes here are therefore not considered breaking.

FerriteProblems.FEBuffer — Type

FEBuffer(K,x,r,f,xold,cellbuffer,state,old_state,time,old_time)

A buffer to hold all values that are required to simulate, but that are uniqely defined from the simulation definition

FerriteProblems.getcellbuffer — Function

FerriteProblems.getcellbuffer(p::FerriteProblem)

Get the cell buffers used during the assembly.

FerriteProblems.get_tolerance_scaling — Function

FerriteProblems.get_tolerance_scaling(p::FerriteProblem)

Get the TolScaling type that controls the convergence measure to be compared with the solver's tolerance

FerriteProblems.update_unknowns! — Function

FerriteProblems.update_unknowns!(p::FerriteProblem)

Update the vector of "old" unknowns to the values of the current vector of unknowns

FerriteProblems.copy_states! — Function

FerriteProblems.copy_states!(to, from)

Perform a "deepcopy!" of states in from into to.

FerriteProblems.update_states! — Function

FerriteProblems.update_states!(p::FerriteProblem)

Update the "old" state variables to the current values. Called after convergence

FerriteProblems.reset_states! — Function

FerriteProblems.reset_states!(p::FerriteProblem)

Reset the current state variables to the old state values. Called after each solution iteration unless the solution has converged.

FerriteProblems.settime! — Function

FerriteProblems.settime!(p::FerriteProblem, new_time)

Set the current time to new_time Called when starting a new step (or when attempting the same step number with a new time increment)

FerriteProblems.update_time! — Function

FerriteProblems.update_time!(p::FerriteProblem)

Update the old time to the current time. Called after convergence

FerriteProblems._create_states — Function

_create_states(dh::AbstractDofHandler, material, cellvalues, initial_conditions)

Create the state variables by calling FerriteAssembly.create_states after calculating the degree of freedom values with Ferrite.jl's applyanalytical! and the `initialconditions` input.

FerriteProblems.filepath — Function

filepath(io::FerriteIO, args...) = joinpath(io.folder, args...)

Get the path of a file relative io's folder

FerriteProblems.datafilepath — Function

datafilepath(io::FerriteIO, num=length(io.datafiles))

Get the path of the data file number num in io.datafiles

FerriteProblems.new_file! — Function

new_file!

FerriteProblems.new_file_if_needed! — Function

new_file_if_needed!

FerriteProblems.update_currentsize! — Function

update_currentsize!

FerriteProblems.getfilenumber — Function

getfilenumber

FerriteProblems.open_if_needed! — Function

open_if_needed!

FerriteProblems.close_io — Function

close_io(io::FerriteIO, post)

Close the currently open file in io, then the postprocessing struct to a jld2 file (if not post != nothing), before finally saving the current io object to a .jld2 file

FerriteProblems.checkkey — Function

checkkey

FerriteProblems.savedata! — Function

savedata!

FerriteProblems.getdata — Function

getdata

FerriteProblems.get_dof2node — Function

get_dof2node

FerriteProblems.dothreaded — Function

dothreaded(def::FEDefinition)

Trait-based dispatch to determine if threaded assembly should be used. Returns a Val{t::Bool}.